requirements.txt
pyautogui
pillow
pytesseract
openai
pytest